Updated version number v0023
Made larva more ignored, due to ability to load build orders that have larva in them.

Spire is now a more upgradable building. Was causing issues when I was trying to get +3+3 flyer armor/attacks.

Arbitrary increase in score when satisfaction occurs. If you created a build previously that timewise was really close to satisfaction, it may choose to not be satisfied for something else it wants, because it doesn't see the bonus score as worth it.

German text now contains some korean.

que?